fix referendum benchmark test
